Explicitly use package_data over include_package_data
Building the source or binary distributions of this python library produces a package without any of the table files present. `include_package_data=True` is insufficient and seemed to not pick up the files.

Indeed, this StackOverflow entry seems to detail the woe: https://stackoverflow.com/questions/7522250/how-to-include-package-data-with-setuptools-distutils

By using `package_data` not only does it now seem to work on both source and binary distributions (ie the files get copied when running `python setup.py sdist bdist`) but there's also finer-grained control on what files to include in the final package/library.